@@ -21,7 +21,7 @@
     "lint": "eslint src --max-warnings 0"
   },
   "peerDependencies": {
-    "react": "^18 || ^19"
+    "react": "^19"
   },
   "devDependencies": {
     "@repo/eslint-config": "*",
@@ -31,11 +31,6 @@
     "@types/react": "^19.1.0",
     "eslint": "^9.27.0",
     "tailwindcss": "^4.1.5",
-    "typescript": "5.8.2",
-    "class-variance-authority": "^0.7.0",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-navigation-menu": "^1.1.4",
-    "@radix-ui/react-separator": "^1.0.3"
+    "typescript": "5.8.2"
   }
 }
