@@ -1,63 +1,12 @@
-const { i18n } = require('./next-i18next.config')
+import type { NextConfig } from "next";
 
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  swcMinify: true,
-  i18n,
-  
-  images: {
-    domains: [
-      'localhost',
-      'everapharm.com',
-      'cdn.everapharm.com',
-      'picsum.photos', // For demo images
-      process.env.NEXT_PUBLIC_CDN_URL ? new URL(process.env.NEXT_PUBLIC_CDN_URL).hostname : '',
-    ].filter(Boolean),
+const nextConfig: NextConfig = {
+  typescript: {
+    ignoreBuildErrors: true,
   },
-
-  async rewrites() {
-    return [
-      {
-        source: '/api/:path*',
-        destination: `${process.env.NEXT_PUBLIC_API_URL || 'http://localhost:4000'}/api/:path*`,
-      },
-    ]
+  eslint: {
+    ignoreDuringBuilds: true,
   },
+};
 
-  async headers() {
-    return [
-      {
-        source: '/:path*',
-        headers: [
-          {
-            key: 'X-DNS-Prefetch-Control',
-            value: 'on',
-          },
-          {
-            key: 'Strict-Transport-Security',
-            value: 'max-age=63072000; includeSubDomains; preload',
-          },
-          {
-            key: 'X-Frame-Options',
-            value: 'SAMEORIGIN',
-          },
-          {
-            key: 'X-Content-Type-Options',
-            value: 'nosniff',
-          },
-          {
-            key: 'Referrer-Policy',
-            value: 'origin-when-cross-origin',
-          },
-        ],
-      },
-    ]
-  },
-
-  experimental: {
-    typedRoutes: true,
-  },
-}
-
-module.exports = nextConfig
+export default nextConfig;
